#10077: Overload binary ge in unary ge
