+++ docs/source/ttnn/ttnn/api.rst
@@ -281,7 +281,6 @@ Pointwise Binary
    ttnn/rsub_bw
    ttnn/bias_gelu_bw
    ttnn/binary_ne_bw
-   ttnn/binary_ge_bw
    ttnn/min_bw
    ttnn/max_bw
    ttnn/div_bw
