--- docs/source/ttnn/ttnn/ttnn/binary_ge_bw.rst
@@ -1,6 +1,0 @@
-.. _ttnn.binary_ge_bw:
-
-ttnn.binary_ge_bw
-#################
-
-.. autofunction:: ttnn.binary_ge_bw
