+++ tests/ttnn/unit_tests/operations/backward/test_backward_ge.py
@@ -16,6 +16,25 @@ from tests.ttnn.unit_tests.operations.backward.utility_funcs import data_gen_wit
         (torch.Size([1, 3, 320, 384])),
     ),
 )
+def test_bw_binary_ge(input_shapes, device):
+    in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
+    grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 100, device)
+
+    tt_output_tensor_on_device = ttnn.ge_bw(grad_tensor, input_tensor, input_tensor)
+    pt_y = torch.zeros_like(grad_data)
+    golden_tensor = [pt_y, pt_y]
+    comp_pass = compare_pcc(tt_output_tensor_on_device, golden_tensor)
+    assert comp_pass
+
+
+@pytest.mark.parametrize(
+    "input_shapes",
+    (
+        (torch.Size([1, 1, 32, 32])),
+        (torch.Size([1, 1, 320, 384])),
+        (torch.Size([1, 3, 320, 384])),
+    ),
+)
 @pytest.mark.parametrize("other", [1.0])
 def test_bw_unary_ge(input_shapes, other, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
